hidden_stores/requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
pillow==11.0.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
pymongo==4.10.1
python-decouple==3.8
reportlab==4.2.5
sqlparse==0.5.1
tzdata==2024.2
    